pyproject.toml
[project]
name = "stock-summarizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "feedparser>=6.0.12",
    "finance-datareader>=0.9.102",
    "googlenews>=1.6.15",
    "lxml>=6.0.2",
    "openai>=2.17.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.54.0",
]
